==> Scripts/00-packages.R <==
#commonly used packages
library(krsp)
library(devtools)
library(data.table)
library(dplyr)
library(lubridate)
library(sf)
library(ggplot2)
library(ggmap)
library(hms)
library(sp)
library(tidyverse)
library(tidyr)
library(concaveman)
library(adehabitatHR)
library(amt)
library(raster)
library(arm)
library(rsq)
library(stringr)
library(effsize)
library(stats)
library(ggeffects)
library(MASS)
library(vcd)
library(gamlss)
library(quantreg)
library(mgcv)
library(jtools)
library(pROC)
library(interactions)
library(ggforce)
library(dunn.test)
library(effectsize)
library(knitr)
library(lme4)
library(lmerTest)
library(broom.mixed)
library(MuMIn)
library(ggeffects)
library(nnet)
library(car)
library(boot)
library(simr)
library(pwr)
library(betareg)
library(glmmTMB)
library(DHARMa)
library(influence.ME)
library(caret)
library(emmeans)
library(glmmTMB)
library(mclogit)
library(scales)
